+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
+* `FIX` Using variables to retrieve array elements always prompts the first element type
 
 ## 3.12.0
 `2024-10-30`