+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
+* `FIX` Improve type narrow with **literal alias type** during completion and signature help
 
 ## 3.12.0
 `2024-10-30`