+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
-* `FIX` Improve type narrow with **literal alias type** during completion and signature help
+
+## 3.13.0
+`2024-11-13`
 * `NEW` Setting: `Lua.type.inferTableSize`: A Small Table array can be infered
 * `NEW` Add custom repository support for addonManager. New configuration setting: `Lua.addonManager.repositoryBranch` and `Lua.addonManager.repositoryPath`
 * `NEW` Infer function parameter types when the function is used as an callback argument and that argument has a `fun()` annotation. Enable with `Lua.type.inferParamType` setting. [#2695](https://github.com/LuaLS/lua-language-server/pull/2695)
@@ -23,6 +25,7 @@
   ---@type int32
   local a = n
   ```
+* `FIX` Improve type narrow with **literal alias type** during completion and signature help
 
 ## 3.12.0
 `2024-10-30`